refactor: Add new managed hooking class implementations
* Part of converting the older code to reduce code re-use and try to make the hooking classes it more extensible
